rename private headers
