+++ include/aws/io/io.h
@@ -15,6 +15,19 @@ AWS_PUSH_SANE_WARNING_LEVEL
 #define AWS_C_IO_PACKAGE_ID 1
 
 struct aws_io_handle;
+typedef void aws_io_set_queue_on_handle_fn(struct aws_io_handle *handle, void *queue);
+typedef void aws_io_clear_queue_on_handle_fn(struct aws_io_handle *handle);
+
+struct aws_io_handle {
+    union {
+        int fd;
+        /* on Apple systems, handle is of type nw_connection_t. On Windows, it's a SOCKET handle. */
+        void *handle;
+    } data;
+    void *additional_data;
+    aws_io_set_queue_on_handle_fn *set_queue;
+    aws_io_clear_queue_on_handle_fn *clear_queue;
+};
 
 enum aws_io_message_type {
     AWS_IO_MESSAGE_APPLICATION_DATA,
