+++ include/aws/io/socket.h
@@ -7,7 +7,6 @@
 
 #include <aws/io/channel.h>
 #include <aws/io/io.h>
-#include <aws/io/private/event_loop_impl.h>
 
 AWS_PUSH_SANE_WARNING_LEVEL
