+++ source/darwin/dispatch_queue_event_loop.c
@@ -17,7 +17,7 @@
 
 #include <unistd.h>
 
-#include "dispatch_queue.h"
+#include "./dispatch_queue_event_loop_private.h" // private header
 #include <Block.h>
 #include <dispatch/dispatch.h>
 #include <dispatch/queue.h>
